+++ Changelog.md
@@ -13,7 +13,8 @@ Language Features:
 
 Compiler Features:
  * C API (``libsolc`` / raw ``soljson.js``): Introduce ``solidity_free`` method which releases all internal buffers to save memory.
-
+ * Commandline interface: Adds new option ``--new-reporter`` for improved diagnostics formatting
+   along with ``--color`` and ``--no-color`` for colorized output to be forced (or explicitly disabled).
 
 Bugfixes: